+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Metas/AIMetas/ChaseBehaviourMeta.cs
@@ -1,0 +1,10 @@
+﻿using GoldSprite.UFsm;
+using System;
+using UnityEngine;
+
+namespace GoldSprite.EntitySystem2D {
+    [CreateAssetMenu]
+    public class ChaseBehaviourMeta : EntityBehaviourMeta {
+        public override Type StateType=> typeof(ChaseState);
+    }
+}