+++ Assets/Packages/UnityPlugins/UFsm/Scripts/Runtime/Fsm/AIs/States/AIState.cs
@@ -7,13 +7,5 @@ namespace GoldSprite.UFsm {
         public new IAIProps Props => Fsm.Props;
 
         public AIState(AIFsm fsm) : base(fsm) { Fsm = fsm; }
-
-        public override string ToString()
-        {
-            return $"[{Props.Name}-{GetType().Name}]";
-        }
-
-        public override bool Enter() => false;
-        public override bool Exit() => true;
     }
 }